# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Closechain AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subcontractors:
    get:
      operationId: listAllSubcontractors
      tags: [subcontractors]
      summary: List all subcontractors across all user projects
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: All subcontractors with project info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubcontractorAcrossProjects"
  /projects/{projectId}/subcontractors:
    get:
      operationId: listSubcontractors
      tags: [subcontractors]
      summary: List subcontractors for a project
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of subcontractors with document counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubcontractorWithProgress"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createSubcontractor
      tags: [subcontractors]
      summary: Add a subcontractor to a project
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubcontractorBody"
      responses:
        "201":
          description: Subcontractor created with auto-assigned documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubcontractorWithProgress"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/{projectId}/subcontractors/import:
    post:
      operationId: importSubcontractors
      tags: [subcontractors]
      summary: Import subcontractors from CSV data
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportSubcontractorsBody"
      responses:
        "201":
          description: Subcontractors imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportSubcontractorsResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/{projectId}/subcontractors/{subcontractorId}:
    delete:
      operationId: deleteSubcontractor
      tags: [subcontractors]
      summary: Remove a subcontractor from a project
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: subcontractorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SubcontractorAcrossProjects:
      type: object
      required: [id, vendorName, vendorCode, csiCode, csiDivision, projectId, projectName, totalDocuments, uploadedDocuments, approvedDocuments, progress]
      properties:
        id:
          type: integer
        vendorName:
          type: string
        vendorCode:
          type: string
        csiCode:
          type: string
        csiDivision:
          type: string
        projectId:
          type: integer
        projectName:
          type: string
        totalDocuments:
          type: integer
        uploadedDocuments:
          type: integer
        approvedDocuments:
          type: integer
        progress:
          type: number
...
    SubcontractorWithProgress:
      type: object
      required: [id, vendorName, vendorCode, csiCode, csiDivision, totalDocuments, uploadedDocuments, approvedDocuments, progress]
      properties:
        id:
          type: integer
        vendorName:
          type: string
        vendorCode:
          type: string
        csiCode:
          type: string
        csiDivision:
          type: string
        totalDocuments:
          type: integer
        uploadedDocuments:
          type: integer
        approvedDocuments:
          type: integer
        progress:
          type: number
    CreateSubcontractorBody:
      type: object
      required: [vendorName, vendorCode, csiCode]
      properties:
        vendorName:
          type: string
          minLength: 1
        vendorCode:
          type: string
          minLength: 1
        csiCode:
          type: string
          minLength: 1
    ImportSubcontractorsBody:
      type: object
      required: [subcontractors]
      properties:
        subcontractors:
          type: array
          items:
            $ref: "#/components/schemas/CreateSubcontractorBody"
    ImportSubcontractorsResponse:
      type: object
      required: [imported, total]
      properties:
        imported:
          type: integer
        total:
          type: integer
        subcontractors:
          type: array
          items:
            $ref: "#/components/schemas/SubcontractorWithProgress"
...